requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=2.0.0
requests>=2.28.0
PyQt5>=5.15.0
matplotlib>=3.7.0
reportlab>=3.6.0
chartpress>=1.3.0